Everything but timeWeights seem to work

+++ src/main/java/dk/itu/map/structures/ArrayLists/ByteArrayList.java
@@ -75,7 +75,7 @@ public class ByteArrayList extends PrimitiveArrayList implements WriteAble{
     public void write(DataOutputStream stream) throws IOException {
         stream.writeInt(size);
         for (int i = 0; i < size; i++) {
-            stream.writeInt(array[i]);
+            stream.writeByte(array[i]);
         }
     }
